+++ R/aggregate.R
@@ -2,20 +2,21 @@
 #'
 #' aggregate an \code{sf} object, possibly union-ing geometries
 #' @param x object of class \link{sf}
-#' @param by see \link[stats]{aggregate}
+#' @param by (see \link[stats]{aggregate}): a list of grouping elements, each as long as the variables in the data frame x. The elements are coerced to factors before use.
 #' @param FUN function passed on to \link[stats]{aggregate}, in case \code{ids} was specified and attributes need to be grouped
 #' @param ... arguments passed on to \code{FUN}
-#' @param union logical; should grouped geometries be unioned using \link{st_union}? 
+#' @param do_union logical; should grouped geometries be unioned using \link{st_union}? 
 #' @param simplify logical; see \link[stats]{aggregate}
-#' @return an \code{sf} object with aggregated attributes and geometries, with an additional grouping variable called \code{Group.1}.
+#' @return an \code{sf} object with aggregated attributes and geometries, with additional grouping variables having the names of \code{names(ids)} or named \code{Group.i} for \code{ids[[i]]}; see the \code{data.frame} method of \link[stats]{aggregate}.
+#' @aliases aggregate
 #' @export
-aggregate.sf = function(x, by, FUN, ..., union = TRUE, simplify = TRUE) {
+aggregate.sf = function(x, by, FUN, ..., do_union = TRUE, simplify = TRUE) {
 
 	crs = st_crs(x)
 	lst = lapply(split(st_geometry(x), by), function(y) do.call(c, y))
 	geom = do.call(st_sfc, lst[!sapply(lst, is.null)])
 
-	if (union)
+	if (do_union)
 		geom = st_union(geom, by_feature = TRUE)
 	st_geometry(x) = NULL
 	x = aggregate(x, by, FUN, ..., simplify = simplify)