+++ src/wkb.cpp
@@ -20,10 +20,26 @@
 #define EWKB_M_BIT    0x40000000
 #define EWKB_SRID_BIT 0x20000000
 
-Rcpp::List read_data(const unsigned char **pt, bool EWKB, int endian, bool addclass, 
-	int *type, uint32_t *srid);
+typedef struct {
+	const unsigned char *pt;
+	size_t size;
+} wkb_buf;
+
+void wkb_read(wkb_buf *wkb, void *dst, int n);
+
+Rcpp::List read_data(wkb_buf *wkb, bool EWKB, bool spatialite, int endian, 
+	bool addclass, int *type, uint32_t *srid);
 void write_data(std::ostringstream& os, Rcpp::List sfc, int i, bool EWKB, 
 		int endian, const char *cls, const char *dim, double prec, int srid);
+
+void wkb_read(wkb_buf *wkb, void *dst, int n) {
+	if (n > wkb->size)
+		throw std::range_error("range check error: WKB buffer too small. Input file corrupt?");
+	if (dst != NULL)
+		memcpy(dst, wkb->pt, n);
+	wkb->pt += n;
+	wkb->size -= n;
+}
 
 // https://stackoverflow.com/questions/105252/how-do-i-convert-between-big-endian-and-little-endian-values-in-c
 template <typename T>
@@ -89,16 +105,72 @@ Rcpp::CharacterVector CPL_raw_to_hex(Rcpp::RawVector raw) {
 	return Rcpp::CharacterVector::create(os.str());
 }
 
-Rcpp::NumericMatrix read_multipoint(const unsigned char **pt, int n_dims, bool swap, 
-		bool EWKB = 0, int endian = 0, Rcpp::CharacterVector cls = "", bool *empty = NULL) {
+void read_spatialite_header(wkb_buf *wkb, uint32_t *srid, bool swap) {
+	// we're at byte 3 now:
+	wkb_read(wkb, srid, sizeof(uint32_t));
+
+	if (swap)
+		*srid = swap_endian<uint32_t>(*srid);
+
+	wkb_read(wkb, NULL, 32); // skip header
+	// verify special marker; if not there, raise error:
+	unsigned char marker;
+	wkb_read(wkb, &marker, 1); // skip header
+	if (marker != 0x7c) { 
+		Rcpp::Rcout << "byte 39 should be 0x7c, but is " << marker << std::endl; // #nocov
+		throw std::range_error("invalid spatialite header"); // #nocov
+	}
+}
+
+void read_gpkg_header(wkb_buf *wkb, uint32_t *srid, int endian) {
+	// http://www.geopackage.org/spec/#gpb_format
+	wkb_read(wkb, NULL, 3); // 'G', 'P', version
+
+	// read flag:
+	unsigned char flag;
+	wkb_read(wkb, &flag, 1);
+	bool swap = ((flag & 0x01) != (int) endian); // endian check
+
+	// read srid, if needed, swap:
+	wkb_read(wkb, srid, 4);
+	if (swap)
+		*srid = swap_endian<uint32_t>(*srid);
+
+	// how much header is there to skip? bbox: 4, 6, 6, or 8 doubles:
+	flag = (flag >> 1) & 0x07; // get bytes 3,2,1
+	int n = 0;
+	if (flag == 1) // [minx, maxx, miny, maxy]
+		n = 32;
+	else if (flag == 2 || flag == 3) // #nocov start
+			// [minx, maxx, miny, maxy, minz, maxz] or [minx, maxx, miny, maxy, minm, maxm]
+		n = 48;
+	else if (flag == 4) // [minx, maxx, miny, maxy, minz, maxz, minm, maxm]
+		n = 64; // #nocov end
+	wkb_read(wkb, NULL, n);
+}
+
+Rcpp::NumericMatrix read_multipoint(wkb_buf *wkb, int n_dims, bool swap,
+		bool EWKB = 0, bool spatialite = false, int endian = 0, Rcpp::CharacterVector cls = "",
+		bool *empty = NULL) {
+
 	uint32_t npts;
-	memcpy(&npts, *pt, sizeof(uint32_t));
+	wkb_read(wkb, &npts, 4);
+
 	if (swap)
 		npts = swap_endian<uint32_t>(npts);
-	(*pt) += 4;
+
 	Rcpp::NumericMatrix ret(npts, n_dims);
 	for (size_t i = 0; i < npts; i++) {
-		Rcpp::List lst = read_data(pt, EWKB, endian, false, NULL, NULL);
+		if (spatialite) {
+			// verify special marker; if not there, raise error:
+			unsigned char marker;
+			wkb_read(wkb, &marker, 1); // absorb the 0x69
+			if (marker != 0x69) {  // #nocov start
+				Rcpp::Rcout << "0x69 marker missing before ring " << i+1 << std::endl;
+				throw std::range_error("invalid spatialite header");
+			} // #nocov end
+		}
+		Rcpp::List lst = read_data(wkb, EWKB, spatialite, endian, false, NULL, NULL);
 		Rcpp::NumericVector vec = lst[0];
 		for (int j = 0; j < n_dims; j++)
 			ret(i,j) = vec(j);
@@ -110,16 +182,30 @@ Rcpp::NumericMatrix read_multipoint(const unsigned char **pt, int n_dims, bool s
 	return ret;
 }
 
-Rcpp::List read_geometrycollection(const unsigned char **pt, int n_dims, bool swap, bool EWKB = 0, 
-		int endian = 0, Rcpp::CharacterVector cls = "", bool isGC = true, bool *empty = NULL) {
+Rcpp::List read_geometrycollection(wkb_buf *wkb, int n_dims, bool swap, bool EWKB = 0, 
+		bool spatialite = false, int endian = 0, Rcpp::CharacterVector cls = "", bool isGC = true, 
+		bool *empty = NULL) {
+
 	uint32_t nlst;
-	memcpy(&nlst, *pt, sizeof(uint32_t));
+	wkb_read(wkb, &nlst, 4);
+
 	if (swap)
 		nlst = swap_endian<uint32_t>(nlst);
-	(*pt) += 4;
+
 	Rcpp::List ret(nlst);
-	for (size_t i = 0; i < nlst; i++)
-		ret[i] = read_data(pt, EWKB, endian, isGC, NULL, NULL)[0];
+
+	for (size_t i = 0; i < nlst; i++) {
+		if (spatialite) {
+			// verify special marker; if not there, raise error
+			unsigned char marker;
+			wkb_read(wkb, &marker, 1); // absorb the 0x69
+			if (marker != 0x69) { // #nocov start
+				Rcpp::Rcout << "0x69 marker missing before ring " << i+1 << std::endl;
+				throw std::range_error("invalid spatialite header");
+			} // #nocov end
+		}
+		ret[i] = read_data(wkb, EWKB, spatialite, endian, isGC, NULL, NULL)[0];
+	}
 	if (cls.size() == 3)
 		ret.attr("class") = cls;
 	if (empty != NULL)
@@ -127,40 +213,40 @@ Rcpp::List read_geometrycollection(const unsigned char **pt, int n_dims, bool sw
 	return ret;
 }
 
-Rcpp::NumericVector read_numeric_vector(const unsigned char **pt, int n, bool swap,
+Rcpp::NumericVector read_numeric_vector(wkb_buf *wkb, int n, bool swap,
 		Rcpp::CharacterVector cls = "") {
 	Rcpp::NumericVector ret(n);
 	for (int i = 0; i < n; i++) {
 		double d;
-		memcpy(&d, *pt, sizeof(double));
+		wkb_read(wkb, &d, 8);
 		if (swap)
 			ret(i) = swap_endian<double>(d);
 		else
 			ret(i) = d;
-		(*pt) += 8;
 	}
 	if (cls.size() == 3)
 		ret.attr("class") = cls;
 	return ret;
 }
 
-Rcpp::NumericMatrix read_numeric_matrix(const unsigned char **pt, int n_dims, bool swap,
+Rcpp::NumericMatrix read_numeric_matrix(wkb_buf *wkb, int n_dims, bool swap,
 		Rcpp::CharacterVector cls = "", bool *empty = NULL) {
+
 	uint32_t npts;
-	memcpy(&npts, *pt, sizeof(uint32_t));
+	wkb_read(wkb, &npts, 4);
+
 	if (swap)
 		npts = swap_endian<uint32_t>(npts);
-	(*pt) += 4;
+
 	Rcpp::NumericMatrix ret(npts, n_dims);
 	for (size_t i = 0; i < npts; i++)
 		for (int j = 0; j< n_dims; j++) {
 			double d;
-			memcpy(&d, *pt, sizeof(double));
+			wkb_read(wkb, &d, 8);
 			if (swap)
 				ret(i, j) = swap_endian<double>(d);
 			else
 				ret(i, j) = d;
-			(*pt) += 8;
 		}
 	if (cls.size() == 3)
 		ret.attr("class") = cls;
@@ -169,16 +255,19 @@ Rcpp::NumericMatrix read_numeric_matrix(const unsigned char **pt, int n_dims, bo
 	return ret;
 }
 
-Rcpp::List read_matrix_list(const unsigned char **pt, int n_dims, bool swap, 
+Rcpp::List read_matrix_list(wkb_buf *wkb, int n_dims, bool swap, 
 		Rcpp::CharacterVector cls = "", bool *empty = NULL) {
+
 	uint32_t nlst;
-	memcpy(&nlst, *pt, sizeof(uint32_t));
+	wkb_read(wkb, &nlst, 4);
+
 	if (swap)
 		nlst = swap_endian<uint32_t>(nlst);
-	(*pt) += 4;
+
 	Rcpp::List ret(nlst);
 	for (size_t i = 0; i < nlst; i++)
-		ret[i] = read_numeric_matrix(pt, n_dims, swap, "");
+		ret[i] = read_numeric_matrix(wkb, n_dims, swap, "");
+
 	if (cls.size() == 3)
 		ret.attr("class") = cls;
 	if (empty != NULL)
@@ -186,26 +275,48 @@ Rcpp::List read_matrix_list(const unsigned char **pt, int n_dims, bool swap,
 	return ret;
 }
 
-Rcpp::List read_data(const unsigned char **pt, bool EWKB = false, int endian = 0, 
-		bool addclass = true, int *type = NULL, uint32_t *srid = NULL) {
+Rcpp::List read_data(wkb_buf *wkb, bool EWKB = false, bool spatialite = false,
+		int endian = 0, bool addclass = true, int *type = NULL, uint32_t *srid = NULL) {
 /*
  pt: handle to the memory buffer
  EWKB: should we read EWKB, as opposed to ISO WKB?
  endian: 0 or 1, indicating big (0) or little (1) endian of the buffer
  addclass: write class information to object?
  type: IF NOT NULL: output the geometry type of object read
- srid: IF NOT NULL: output the srid read
- */
+ srid: IF NOT NULL: output the srid read; NULL indicates a nested call
+*/
 
 	Rcpp::List output(1); // to deal with varying result type
-	bool swap = ((int) (**pt) != (int) endian); // endian check
-	(*pt)++;
+
+	if (srid != NULL && wkb->size > 2 && wkb->pt[0] == 'G' && wkb->pt[1] == 'P') // GPKG header? skip:
+		read_gpkg_header(wkb, srid, endian);
+
+	if (spatialite && srid != NULL)
+		wkb_read(wkb, NULL, 1); // starting 0x00 contains no information
+
+	unsigned char swap_char;
+	bool swap;
+
+	if (spatialite && srid == NULL) // nested call: don't read swap:
+		swap = false;
+	else {
+		wkb_read(wkb, &swap_char, 1);
+		swap = ((int) swap_char != (int) endian); // endian check
+	}
+	if (spatialite) {
+		if (swap) 
+			throw std::range_error("reading non-native endian spatialite geometries not supported");
+		if (srid != NULL) // not nested:
+			read_spatialite_header(wkb, srid, swap);
+	}
+	
 	// read type:
 	uint32_t wkbType;
-	memcpy(&wkbType, *pt, sizeof(uint32_t));
+	wkb_read(wkb, &wkbType, 4);
+
 	if (swap)
 		wkbType = swap_endian<uint32_t>(wkbType);
-	(*pt) += 4;
+
 	int sf_type = 0, n_dims = 0;
 	std::string dim_str = ""; 
 	if (EWKB) { // EWKB: PostGIS default
@@ -224,11 +335,10 @@ Rcpp::List read_data(const unsigned char **pt, bool EWKB = false, int endian = 0
 			dim_str = "XYZM";
 		if (wkbSRID != 0) {
 			if (srid != NULL) { 
-				memcpy(srid, *pt, sizeof(uint32_t));
+				wkb_read(wkb, srid, 4);
 				if (swap)
 					*srid = swap_endian<uint32_t>(*srid);
 			}
-			(*pt) += 4;
 		}
 	} else { // ISO
 		sf_type = wkbType % 1000;
@@ -238,78 +348,79 @@ Rcpp::List read_data(const unsigned char **pt, bool EWKB = false, int endian = 0
 			case 2: n_dims = 3; dim_str = "XYM"; break; 
 			case 3: n_dims = 4; dim_str = "XYZM"; break; 
 			default:
-				throw std::range_error("unknown wkbType dim in switch");
+				Rcpp::Rcout << "wkbType: " << wkbType << std::endl;
+				throw std::range_error("unsupported wkbType dim in switch");
 		}
 	}
 	bool empty = false;
 	switch(sf_type) {
 		case SF_Point: 
-			output[0] = read_numeric_vector(pt, n_dims, swap, addclass ?
+			output[0] = read_numeric_vector(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "POINT", "sfg") : "");
 			break;
 		case SF_LineString:
-			output[0] = read_numeric_matrix(pt, n_dims, swap, addclass ?
+			output[0] = read_numeric_matrix(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "LINESTRING", "sfg") : "", &empty);
 			break;
 		case SF_Polygon: 
-			output[0] = read_matrix_list(pt, n_dims, swap, addclass ?
+			output[0] = read_matrix_list(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "POLYGON", "sfg") : "", &empty);
 			break;
 		case SF_MultiPoint: 
-			output[0] = read_multipoint(pt, n_dims, swap, EWKB, endian, addclass ?  
+			output[0] = read_multipoint(wkb, n_dims, swap, EWKB, spatialite, endian, addclass ?  
 				Rcpp::CharacterVector::create(dim_str, "MULTIPOINT", "sfg") : "", &empty); 
 			break;
 		case SF_MultiLineString:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "MULTILINESTRING", "sfg"), false, &empty);
 			break;
 		case SF_MultiPolygon:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "MULTIPOLYGON", "sfg"), false, &empty);
 			break;
 		case SF_GeometryCollection: 
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "GEOMETRYCOLLECTION", "sfg"), true,
 				&empty);
 			break;
 		case SF_CircularString:
-			output[0] = read_numeric_matrix(pt, n_dims, swap, addclass ?
+			output[0] = read_numeric_matrix(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "CIRCULARSTRING", "sfg") : "", &empty);
 			break;
 		case SF_CompoundCurve:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "COMPOUNDCURVE", "sfg"), true, &empty); 
 			break;
 		case SF_CurvePolygon:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "CURVEPOLYGON", "sfg"), true, &empty); 
 			break;
 		case SF_MultiCurve:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "MULTICURVE", "sfg"), true, &empty);
 			break;
 		case SF_MultiSurface:
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "MULTISURFACE", "sfg"), true, &empty);
 			break;
 		case SF_Curve:
-			output[0] = read_numeric_matrix(pt, n_dims, swap, addclass ?
+			output[0] = read_numeric_matrix(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "CURVE", "sfg") : "", &empty); 
 			break;
 		case SF_Surface: 
-			output[0] = read_matrix_list(pt, n_dims, swap, addclass ?
+			output[0] = read_matrix_list(wkb, n_dims, swap, addclass ?
 				Rcpp::CharacterVector::create(dim_str, "SURFACE", "sfg") : "", &empty);
 			break;
 		case SF_PolyhedralSurface: 
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "POLYHEDRALSURFACE", "sfg"), false, &empty);
 			break;
 		case SF_TIN: 
-			output[0] = read_geometrycollection(pt, n_dims, swap, EWKB, endian,
+			output[0] = read_geometrycollection(wkb, n_dims, swap, EWKB, spatialite, endian,
 				Rcpp::CharacterVector::create(dim_str, "TIN", "sfg"), false, &empty);
 			break;
 		case SF_Triangle:
-			output[0] = read_matrix_list(pt, n_dims, swap,
+			output[0] = read_matrix_list(wkb, n_dims, swap,
 				Rcpp::CharacterVector::create(dim_str, "TRIANGLE", "sfg"), &empty);
 			break;
 		default: {
@@ -319,7 +430,7 @@ Rcpp::List read_data(const unsigned char **pt, bool EWKB = false, int endian = 0
 	}
 	if (type != NULL) {
 		if (empty)
-			*type = 0;
+			*type = -sf_type;
 		else
 			*type = sf_type;
 	}
@@ -327,7 +438,7 @@ Rcpp::List read_data(const unsigned char **pt, bool EWKB = false, int endian = 0
 }
 
 // [[Rcpp::export]]
-Rcpp::List CPL_read_wkb(Rcpp::List wkb_list, bool EWKB = false, int endian = 0) {
+Rcpp::List CPL_read_wkb(Rcpp::List wkb_list, bool EWKB = false, bool spatialite = false, int endian = 0) {
 	Rcpp::List output(wkb_list.size());
 
 	int type = 0, last_type = 0, n_types = 0, n_empty = 0;
@@ -336,10 +447,16 @@ Rcpp::List CPL_read_wkb(Rcpp::List wkb_list, bool EWKB = false, int endian = 0) 
 	for (int i = 0; i < wkb_list.size(); i++) {
 		Rcpp::checkUserInterrupt();
 		Rcpp::RawVector raw = wkb_list[i];
-		const unsigned char *pt = &(raw[0]);
-		output[i] = read_data(&pt, EWKB, endian, true, &type, &srid)[0];
-		if (type == 0)
+		wkb_buf wkb;
+		wkb.pt = &(raw[0]);
+		wkb.size = raw.size();
+		// const unsigned char *pt = &(raw[0]);
+		output[i] = read_data(&wkb, EWKB, spatialite, endian, true, &type, &srid)[0];
+		if (type <= 0) {
+			type = -type;
 			n_empty++;
+		}
+		// Rcpp::Rcout << "type is " << type << "\n";
 		if (n_types <= 1 && type != last_type) {
 			last_type = type;
 			n_types++; // check if there's more than 1 type:
@@ -347,7 +464,7 @@ Rcpp::List CPL_read_wkb(Rcpp::List wkb_list, bool EWKB = false, int endian = 0) 
 	}
 	output.attr("single_type") = n_types <= 1; // if 0, we have only empty geometrycollections
 	output.attr("n_empty") = (int) n_empty;
-	if (EWKB == true && srid != 0)
+	if ((EWKB || spatialite) && srid != 0)
 		output.attr("srid") = (int) srid;
 	return output;
 }